fix btapp crash when disable BT with BR/EDR/BLE HID connected.
cast the hh_poll_thread_id type to int, thus we could compare it to 0 normally.

+++ btif/co/bta_hh_co.c
@@ -403,7 +403,7 @@ static inline void btif_hh_close_poll_thread(btif_hh_device_t *p_dev)
 {
     APPL_TRACE_DEBUG("%s", __FUNCTION__);
     p_dev->hh_keep_polling = 0;
-    if(p_dev->hh_poll_thread_id > 0)
+    if((int)(p_dev->hh_poll_thread_id) > 0)
         pthread_join(p_dev->hh_poll_thread_id,NULL);
 
     return;
